Rename every occurrences of MMIX to MIX
So I searched MMIX on internet and everything looked very different to what I saw in my book. It seems I was working with MIX and there's a good chance I just misread the name like an idiot. Can't wait to get back to the book to know for sure. Happy new year!